Setup a planted pattern model

Supplied with a dataframe of planted patterns of b-nodes, will return all the neccesary components to draw from SBM model with set planted structure using draw_from_model().

setup_planted_pattern_model(planted_patterns, num_noise_nodes = 10,
  noise_p = 0.2, planted_p_on = 0.95, planted_p_off = 0.05)

Arguments

planted_patterns
Tibble containing planted patterns for node a groups by activating b nodes with 1 or 0 in columns prefixed with pattern b1, b2, .... Additionally a column size is used to control how many a nodes have the planted pattern.

num_noise_nodes
How many additional b nodes with no pattern are added to model?

noise_p
The individual lambda for connections between the noise b-nodes and any a-node.

planted_p_on
The lambda for a planted pattern b-node that has been turned on in a pattern.

planted_p_off
The lambda for a planted pattern b-node that has been turned off in pattern.

Value

List with group membership vectors for both node types (b_a, b_b), and the Lambda matrix for edge behavior (Lambda).

Examples

my_patterns <- dplyr::tribble(
  ~b1,  ~b2,  ~b3,  ~b4,  ~size,
    1,    1,    0,    0,     10,
    1,    1,    1,    0,     15,
    0,    0,    0,    1,     13,
    0,    0,    1,    1,     24
)
setup_planted_pattern_model(my_patterns, num_noise_nodes = 15)

#> $b_a
#>  [1] 1 1 1 1 1 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 3 3 3
#> [39] 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4
#> 
#> $b_b
#>  [1] 1 2 3 4 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5
#> 
#> $Lambda
#> # A tibble: 20 x 3
#>        a     b avg_num_cons
#>    <int> <int>        <dbl>
#>  1     1     1         0.95
#>  2     2     1         0.95
#>  3     3     1         0.05
#>  4     4     1         0.05
#>  5     1     2         0.95
#>  6     2     2         0.95
#>  7     3     2         0.05
#>  8     4     2         0.05
#>  9     1     3         0.05
#> 10     2     3         0.95
#> 11     3     3         0.05
#> 12     4     3         0.95
#> 13     1     4         0.05
#> 14     2     4         0.05
#> 15     3     4         0.95
#> 16     4     4         0.95
#> 17     1     5         0.2 
#> 18     2     5         0.2 
#> 19     3     5         0.2 
#> 20     4     5         0.2 
#>